initial commit for solving issue #15
the pair of notebooks attached to this commit deomonstrate that the mistral-7b model can be downloaded and used on a system that has at least 16 GB of GPU memory when the model is quantized in 4-bit.

future related work will include building a simple RAG system with this model. work to be completed through a separate issue

this closes issue #15